deps: update brotli to v1.0.9
Refs: https://github.com/google/brotli/releases/tag/v1.0.8
Refs: https://github.com/google/brotli/releases/tag/v1.0.9

PR-URL: https://github.com/nodejs/node/pull/34937

+++ deps/brotli/c/enc/utf8_util.c
@@ -77,7 +77,7 @@ BROTLI_BOOL BrotliIsMostlyUTF8(
     i += bytes_read;
     if (symbol < 0x110000) size_utf8 += bytes_read;
   }
-  return TO_BROTLI_BOOL(size_utf8 > min_fraction * (double)length);
+  return TO_BROTLI_BOOL((double)size_utf8 > min_fraction * (double)length);
 }
 
 #if defined(__cplusplus) || defined(c_plusplus)
